requirements.txt
numpy
matplotlib
nltk
pandas
wordcloud
textblob
ultralytics
opencv-python
pillow
